% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast_18S_reformat}
\alias{blast_18S_reformat}
\title{Process a tabular output from blastn (BLAST+)}
\usage{
blast_18S_reformat(file_name)
}
\arguments{
\item{file_name}{The name of the BLAST file with full path}
}
\value{
TRUE if the function has been successful.

The modified table is saved by changing the name of the file by replacing the extension by \strong{_pr2.tsv}.

The summary table is saved by changing the name of the file by replacing the extension by \strong{_summary.tsv}.
}
\description{
The BLAST file must originate from blastn with the follwing output format option:

-outfmt"6 qseqid sseqid sacc stitle sscinames staxids sskingdoms sblastnames pident
slen length mismatch gapopen qstart qend sstart send evalue bitscore"

It is very important that the columns are in this precise order and no column is missing.\cr

For the formating options see:
\itemize{
\item \href{https://www.ncbi.nlm.nih.gov/books/NBK279684/}{NCBI}
\item \href{http://www.metagenomics.wiki/tools/blast/blastn-output-format-6}{Metagenomics wiki}.
}

What does the function do :
\enumerate{
\item Group all GenBank accession
\item Obtain taxonomy from GenBank (note the GenBank taxonomy is now in the PR2 database after downloading from \url{ftp://ftp.ncbi.nlm.nih.gov/pub/taxonomy/new_taxdump/})
\item Check if the sequence is in PR2 and get the PR2 taxo (this is done with the pr2 database package)
\item Detect whether the Subject sequence is uncultured or not
\item Merge back into the BLAST file
\item Compute a summary with best hit, best hit to PR2, best hit to cultured, taxo consensus (identity>96%), contradiction at division level (identity>90%)
}

The \strong{modified BLAST output file} includes additional columns
\itemize{
\item kingdom -> species : PR2 taxonomy for those accession numbers that are present in PR2
\item hit_rank : the rank of the hit based on decreasing % identity and decreasing bit scores
\item uncultured : TRUE if the hit corresponds to an uncultured item
\item hit_lineage : GenBank taxonomy of the hit
}

The \strong{summary file} contains several set of columns
\enumerate{
\item The top hit (column with prefix hit_top_)
\item The top hit for which a PR2 sequence is available (columns starting with hit_pr2_)
\item The top hit corresponding to a culture or an isolate (columns starting with hit_cul)
\item A "consensus" taxonomy based on all the hits with more than 98\% identity, keeping the most frequent hit at the genus level and using the sum of the bit scores to decide if there are some ties.
\item Contradiction between hits >90\% identity at the division level
}
}
\section{To do}{

}

\section{Column names}{

The columns for the Blast are named as follows. For the summary a prefix is added\preformatted{  hit_top_ / hit_pr2_ / hit_cult_

  query_id, hit_id, hit_acc, hit_title, hit_sci_names

  hit_tax_ids, hit_super_kingdoms, hit_blast_names,

  pct_identity, hit_length, alignment_length, mismatches,

  gap_opens, query_start, query_end, hit_start, hit_end,

  evalue, bit_score
}
}

\section{Programming notes}{

The following functions must be used with libary qualifier \strong{dplyr::} because they are also in the plyr library
\itemize{
\item ungroup
\item desc
\item rename
}

Uses the pr2database package for faster access (much faster !!)
}

\examples{
blast_18S_reformat("C:/BLAST_output.tsv")
}
