% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast_summary}
\alias{blast_summary}
\title{Write a summary for a tabular output from blastn (BLAST+)}
\usage{
blast_summary(file_name)
}
\arguments{
\item{file_name}{The name of the BLAST file with full path}
}
\value{
TRUE if the function has been successful.

The summary table is saved by changing the name of the file by replacing the extension by \strong{_summary.tsv}.
}
\description{
The BLAST file must originate from blastn with the follwing output format option:

-outfmt"6 qseqid sseqid sacc stitle sscinames staxids sskingdoms sblastnames pident
slen length mismatch gapopen qstart qend sstart send evalue bitscore"

It is very important that the columns are in this precise order and no column is missing.\cr

For the formating options see:
\itemize{
\item \href{https://www.ncbi.nlm.nih.gov/books/NBK279684/}{NCBI}
\item \href{http://www.metagenomics.wiki/tools/blast/blastn-output-format-6}{Metagenomics wiki}.
}

What does the function do :
\enumerate{
\item Group all GenBank accession
\item Obtain taxonomy from GenBank (note the GenBank taxonomy is now in the PR2 database after downloading from \url{ftp://ftp.ncbi.nlm.nih.gov/pub/taxonomy/new_taxdump/})
\item Merge back into the BLAST file
\item Compute a summary with best hit,
}

The \strong{summary file} contains several set of columns
\enumerate{
\item The top hit (column with prefix hit_top_)
}
}
\section{To do}{

}

\section{Column names}{

The columns for the Blast are named as follows. For the summary a prefix is added\preformatted{  query_id, hit_id, hit_acc, hit_title, hit_sci_names

  hit_tax_ids, hit_super_kingdoms, hit_blast_names,

  pct_identity, hit_length, alignment_length, mismatches,

  gap_opens, query_start, query_end, hit_start, hit_end,

  evalue, bit_score
}
}

\section{Programming notes}{

The following functions must be used with libary qualifier \strong{dplyr::} because they are also in the plyr library
\itemize{
\item ungroup
\item desc
\item rename
}

Uses the local version of the PR2 database for faster access (much faster !!)
}

\examples{
blast_reformat("C:/BLAST_output.txt")
}
