% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{build_drgus_target_matrix}
\alias{build_drgus_target_matrix}
\title{Builds drugs target association matrix from given drugs_targets dataframe}
\usage{
build_drgus_target_matrix(drugs_targets, drugs_column, targets_column)
}
\arguments{
\item{drugs_targets}{dataframe that contains drugs_targets data}

\item{drugs_column}{name of drugs column in the given dataset}

\item{targets_column}{name of targets column in the given dataset}
}
\value{
dataframe with following sepecifications:
* first column as drugs ids
* targets count other columns (with targets ids as headers)
* values for these columns will be either:
* 1 there is a relation,
* 0 there is no relation.
}
\description{
The dataframe must contains drugs and targets columns.
}
