% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors_euk_class}
\alias{colors_euk_class}
\title{Color for classes.}
\format{structure(colors_euk_class$color_name,.Names=colors_euk_class$class)}
\source{
data-raw/colors.xlsx (update by running  data-raw/data_update.R )
}
\usage{
colors_euk_class
}
\description{
Color for classes.
}
\keyword{datasets}
