% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors_euk_division}
\alias{colors_euk_division}
\title{Color for divisions.}
\format{
structure(colors_euk_division$color_name,.Names=colors_euk_division$division)
}
\source{
data-raw/colors.xlsx (update by running  data-raw/data_update.R )
}
\usage{
colors_euk_division
}
\description{
Color for divisions.
}
\keyword{datasets}
