% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors_euk_genus}
\alias{colors_euk_genus}
\title{Color for genera.}
\format{structure(colors_euk_genus$color_name,.Names=colors_euk_genus$genus)}
\source{
data-raw/colors.xlsx (update by running  data-raw/data_update.R )
}
\usage{
colors_euk_genus
}
\description{
Color for genera.
}
\keyword{datasets}
