% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{dada2_assign_correct}
\alias{dada2_assign_correct}
\title{Correct assignation based on bootstrap values}
\usage{
dada2_assign_correct(df, boot_threshold = 80)
}
\arguments{
\item{df}{Data frame with columns domain:species and domain_boot:species_boot}

\item{boot_threshold}{Boot threshold for correct assignement}
}
\value{
A data frame with the new assignements and columns domain_dada2:species_dada2
}
\description{
Columns with low bootstrap are reassigned to previous column with X added
}
