% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{dada2_export}
\alias{dada2_export}
\title{Export a data frame to dada2 AssignTaxonomy compatible format}
\usage{
dada2_export(df, file_name, taxo_levels_number = 9)
}
\arguments{
\item{df}{data frame - should contain 8 or 9 columns with taxo levels and 1 column with sequences}

\item{file_name}{character - full path of file where to save}
}
\value{
Write the file in compressed format (.gz)
}
\description{
This will save  data frame to dada2 AssignTaxonomy compatible format
}
\examples{
pr2_export_dada2(my_sequences, "C:/Daniel/myfile.fas.gz", 8)
}
