% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_append_records}
\alias{db_append_records}
\title{Write new records to a  table.}
\usage{
db_append_records(database_info, table_name, table_df)
}
\arguments{
\item{table_name}{Character - Name of the table}

\item{table_df}{Dataframe - Must have the same columns than the database table}
}
\value{
TRUE if successful
}
\description{
The data frame must have the same fields than the table (previous name = db_write)
}
\examples{
db_append_records(db_con, "taxo", taxo_new_records)
}
