% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Establish a connection database}
\usage{
db_connect(db_name)
}
\arguments{
\item{db_name}{name of the database ("pr2", "metapr2", "rcc")}
}
\value{
A database connection
}
\description{
The \strong{db_config} must be read before using read_db_config
}
\examples{
db_config <- read_db_config("C:/daniel.vaulot/Databases/google_pc.yaml") # Must always use db_config
pr2_con <- db_connect("pr2")
}
