% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect_sqlite}
\alias{db_connect_sqlite}
\title{Establish a connection to a SQLite database}
\usage{
db_connect_sqlite(file_name)
}
\arguments{
\item{file_name}{Character, name of the file to store the database}
}
\value{
A database connection
}
\description{
Establish a connection to a SQLite database
}
\examples{
pr2_con <- db_connect("PR2.sqlite")
}
