% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect an existing connection}
\usage{
db_disconnect(db_con)
}
\arguments{
\item{db_con}{Database connection}
}
\description{
Disconnect an existing connection
}
\examples{
db_disconnect(pr2_con)
}
