% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_execute_query_vector}
\alias{db_execute_query_vector}
\title{Execute a vector of query.}
\usage{
db_execute_query_vector(database_info, query_vector)
}
\arguments{
\item{query}{Character vector - Queries}
}
\value{
A vector with the number of rows affected by each query
}
\description{
Execute a vector of query.
}
