% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_info}
\alias{db_info}
\title{Get the connection info for a specific database}
\usage{
db_info(database_name,
  file_cnf = "C:/daniel.vaulot@gmail.com/Databases/MySQL/my.cnf")
}
\arguments{
\item{database_name}{One of the following choices "rcc", "pr2_google", "metapr2_google", "rcc_local", "pr2_local", "metapr2_local"}

\item{file_cnf}{by default "C:/daniel.vaulot@gmail.com/Databases/MySQL/my.cnf" but can be moved to another location (e.g. when using dvutils on a server)}
}
\value{
A list with the database connection info "user","password","dbname","host"
}
\description{
To be used with db_connect (see the example).
For the database on Scrol server, one can go though SSH (e.g. when using EduRoam or at NTU).
However at present this connection does not work.
It is first necessary to set up a tunnel through Putty (see https://blog.devolutions.net/2017/4/how-to-configure-an-ssh-tunnel-on-putty).
Same thing for SSL (Google database).
The connection info are now stored in a file (my.cnf). See https://github.com/r-dbi/RMySQL/issues/131
}
\examples{
db_connect(db_info("rcc"))
}
