% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_sql_escape}
\alias{db_sql_escape}
\title{Put a string between single quotes}
\usage{
db_sql_escape(sql_value)
}
\arguments{
\item{sql_value}{String or value to be put between single quotes}
}
\value{
A string between single quotes
}
\description{
Replaces the dbplyr escape function that returns vectors
}
\examples{
db_sql_escape("Eukaryota")
}
