% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_update_field_string}
\alias{db_update_field_string}
\title{Build a string to update a field in a database}
\usage{
db_update_field_string(field_name, field_value, append_comma = TRUE)
}
\arguments{
\item{field_name}{Character}

\item{field_value}{Numeric or string}

\item{append_comma}{TRUE  append a ", " if there more than one field value}
}
\value{
A string in the format "field_name = 'field_value'"
}
\description{
Build a string to update a field in a database
}
\examples{
query <- update_field_string("pr2_accession", "ZZ220100", append_comma=FALSE)
}
