% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs.R
\name{drug_descriptors_categories}
\alias{drug_descriptors_categories}
\title{Returns drugs descriptors categories that can be used to calculate drugs
features.}
\usage{
drug_descriptors_categories()
}
\value{
character vector of available descriptors categories
}
\description{
Each category contains list of descriptors
}
\examples{
\dontrun{
  drug_descriptors_names(drug_descriptors_categories()[1])
}
}
