% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs.R
\name{drug_descriptors_names}
\alias{drug_descriptors_names}
\title{Returns drugs descriptors names that can be used to calculate drugs
features.}
\usage{
drug_descriptors_names(category = "all")
}
\arguments{
\item{category}{character value represents one of
\code{drug_descriptors_categories}}
}
\value{
character vector of given descriptor category names
}
\description{
Returns drugs descriptors names that can be used to calculate drugs
features.
}
\examples{
\dontrun{
  # calculate all available drugs features
  calculate_drugs_AIO(drug_descriptors_names())
}
}
