% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{fasta_filter}
\alias{fasta_filter}
\title{Filter a fasta file}
\usage{
fasta_filter(
  file_name,
  min_length = 100,
  max_length = 10000,
  type = "AA",
  max_ambig = 5
)
}
\arguments{
\item{file_name}{Name the input fasta file}

\item{min_length}{Minimum length of sequences to keep}

\item{max_length}{Maximum length of sequences to keep}

\item{type}{"DNA" or "AA"}

\item{max_ambig}{Maximum number of ambiguites}
}
\value{
TRUE if it terminates OK
}
\description{
Read a fasta file and write back a fasta file keeping only sequences with less than max_ambig and between min and max length.
The file name is changed from xxx.fas to xxx.filtered.fas.
Prints also the number of sequences in and out.
}
\section{To do}{

Nothing =)
}

\examples{
fasta_filter("protein.faa", min_length=100, max_length=10000, type="AA", max_ambig=5)
}
