% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{fasta_read}
\alias{fasta_read}
\title{Read a fasta file into a data frame}
\usage{
fasta_read(file_name, compress = FALSE)
}
\arguments{
\item{file_name}{Character, where to save the fasta file}

\item{compress}{If TRUE reads a gz file}
}
\description{
Read a fasta file from a set of sequences into a data frame
}
\examples{
df <- fasta_read(df,"otu_taxo.fasta", compress=FALSE)
}
