% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{fasta_write}
\alias{fasta_write}
\title{Write a fasta file with the taxonomy}
\usage{
fasta_write(df, file_name, compress = FALSE, taxo_include = TRUE,
  taxo_separator = "|")
}
\arguments{
\item{df}{The data frame with the otu names, the taxonomy and the sequences. It should have the following columns (with exactly these names)\preformatted{  * seq_name : the sequence name
  * supergroup: species
  * sequence
}}

\item{file_name}{Character, where to save the fasta file}

\item{compress}{If TRUE produces a gz file}

\item{taxo_include}{If TRUE then add taxo information which must be provided}

\item{taxo_separator}{Character used to separate the different taxonomic levels
TRUE if it terminates OK}
}
\description{
Write a fasta file from a set of sequences
Option : add to the definition line the the taxonomy separated by separator character (e.g. |)

Otu0001|Alveolata|Dinophyta|Syndiniales|Dino-Group-I|Dino-Group-I-Clade-1|Dino-Group-I-Clade-1_X|Dino-Group-I-Clade-1_X_sp.

AGCTCCAATAGCGTATATTAAAGTTGTTGCGGTTAAAAAGCTCGTAGTTGGA...
}
\examples{
fasta_write(df,"otu_taxo.fasta", compress=FALSE, include_taxo=TRUE, taxo_separator=";")
}
