% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{fastq_subsample}
\alias{fastq_subsample}
\title{Subsample fastq files}
\usage{
fastq_subsample(fastq_path, n_seq = 1000, random = FALSE)
}
\arguments{
\item{fastq_path}{File directories}

\item{n_seq}{Number of sequences to subsample}

\item{random}{If TRUE, then sample a range around n_seq (gaussian with sd=0.1*nseq)}
}
\value{
TRUE if it terminates OK
}
\description{
Subsample fastq files for the first n_seq sequences (! not a random sample)
}
\section{To do}{

Implement compressed files
}

\examples{
fastq_subsample("C:Mydata", n_seq=1000, random=FALSE)
}
