% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{file_unix2dos}
\alias{file_unix2dos}
\title{Convert between Unix and Dos text format}
\usage{
file_unix2dos(filename)
}
\value{
TRUE if successful
}
\description{
Read the Unix text file and writes back to DOS format
}
\examples{
file_unix2dos("my_dos.txt")
}
