% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filename_append}
\alias{filename_append}
\title{Append a string at the end of a file name before the exension}
\usage{
filename_append(filename, new_end)
}
\value{
New file name
}
\description{
The extension is unchanged
}
\examples{
# Will result in "mfile.pr2.txt",
new_name <- filename_append("mfile.txt",".pr2")
}
