% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filename_change_ext}
\alias{filename_change_ext}
\title{Change the extension of the file name}
\usage{
filename_change_ext(filename, new_ext)
}
\value{
New file name
}
\description{
Change the extension of the file name
}
\examples{
# Will result in "mfile.pr2",
new_name <- filename_change_ext("mfile.txt","pr2")
}
