% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.R
\name{genbank_download}
\alias{genbank_download}
\title{Download sequences from GenBank}
\usage{
genbank_download(accession, directory)
}
\arguments{
\item{accession}{Character vector of accession numbers}

\item{directory}{Character, directory name which must end by "/"}
}
\value{
TRUE if successful
}
\description{
Write GenBank files from an vector of accession numbers
}
\examples{
genbank_download(c("JX015376", "JQ768406", "LT621940"), "genbank/")
}
