% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.R
\name{genbank_download_parse_rentrez}
\alias{genbank_download_parse_rentrez}
\title{Download and parse sequences from GenBank using rentrez only}
\usage{
genbank_download_parse_rentrez(accession, sequence_keep = TRUE)
}
\arguments{
\item{accession}{Character vector of accession numbers}

\item{sequence_keep}{Logical, if FALSE the sequence is not returned in the file data frame}
}
\value{
Data frame with the metadata information.
}
\description{
Does not write the GenBank file.
Note: use only sequence_keep = TRUE for simple gene sequences, else it may crash
}
\examples{
gb_metadata <- genbank_download_parse(accession = c("JX015376", "JQ768406", "LT621940"), sequence_keep=TRUE)
}
