% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.R
\name{genbank_features}
\alias{genbank_features}
\title{Read features of sequences from GenBank}
\usage{
genbank_features(accession, directory)
}
\arguments{
\item{accession}{Character vector of accession numbers}

\item{directory}{Character, directory name which must end by "/"}
}
\value{
Data frame with the features information.
}
\description{
Return a dataframe containing all features for a vector of accession numbers. The GenBank files must exist.
}
\examples{
gb_features <- genbank_features(c("JX015376", "JQ768406", "LT621940"), "genbank/")
}
