% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.R
\name{genbank_locus}
\alias{genbank_locus}
\title{Read locus from a set of existing GenBank files}
\usage{
genbank_locus(accession, directory)
}
\arguments{
\item{accession}{Character vector of accession numbers}

\item{directory}{Character, directory name which must end by "/"}
}
\value{
Vector with the locus information.
}
\description{
Return a vector containing the value of a GenBank locus for a vector of accession numbers. The GenBank files must exist
}
\examples{
gb_locus <- genbank_locus(c("JX015376", "JQ768406", "LT621940"), "genbank/")
}
