% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.R
\name{genbank_parse_rentrez_xml}
\alias{genbank_parse_rentrez_xml}
\title{Download and parse sequences from GenBank using rentrez and XML}
\usage{
genbank_parse_rentrez_xml(
  accession,
  acc_step = 50,
  seq_length_max = 15000,
  seq_length_discard = 3e+05
)
}
\arguments{
\item{accession}{Character vector of accession numbers}

\item{acc_step}{Integer, number of accession numbers for each query.}

\item{seq_length_max, }{maximum length of sequences returned (set to zero to prevent returning sequence)}

\item{seq_length_discard, }{if the sequence length is longer than this value, no information is returned.}
}
\value{
Data frame with the metadata information.
}
\description{
Does not write the GenBank file.
Notes:
\itemize{
\item Use only sequence_keep = TRUE for simple gene sequences, else it may crash
\item Use acc_step = 1 if several accession numbers point to the same accession number (case of metagenomes).
In this case the queries are sent one by one and the old accession number is kept.
}
}
\examples{
gb_metadata <- genbank_parse_rentrez_xml(accession = c("JX015376", "JQ768406", "LT621940"), seq_length_max = 15000)
}
