% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.R
\name{genbank_search}
\alias{genbank_search}
\title{Search GenBank}
\usage{
genbank_search(query, db = "nuccore", seq_max = 20000)
}
\arguments{
\item{query}{Character string for ENTREZ search}

\item{db}{Character string for database to be used}

\item{seq_max}{Maximum number of sequences to obtain}
}
\value{
Data frame with columns with information about the sequences
\itemize{
\item genbank_accession
\item gb_definition
\item gb_date
\item gb_organism
\item ...
}
}
\description{
Return sequences summary information from an ENTREZ search query
}
\examples{
genbank_search(query = "28S[TITL] AND rRNA[TITL] AND Chlorophyta[ORGN]", seq_max = 500)
}
