% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.R
\name{genbank_taxonomy}
\alias{genbank_taxonomy}
\title{Read NCBI taxo}
\usage{
genbank_taxonomy(taxo_id)
}
\arguments{
\item{taxo_id}{A vector of genbank taxonomy ids}
}
\value{
A data frame with 2 columns.
}
\description{
Reads directly from NCBI information taxo_lineage from set of taxo_id.

Note : the full NCBI taxonomy has been uploaded in the PR2 database, so this function is not really necessary.
}
\examples{
gb_tax <- genbank_taxonomy(c("1230134", "1230316", "1905175"))
}
