% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{get_primer_position}
\alias{get_primer_position}
\title{Get primer position on sequence set}
\usage{
get_primer_position(
  primer_seq,
  target_seq,
  orientation = "fwd",
  mismatches = 0
)
}
\arguments{
\item{primer_seq}{primer sequence}

\item{target_seq}{target sequences (vector)}

\item{orientation}{"fwd" or "rev"}

\item{mismatches}{Number of mismatches allowed}
}
\value{
data frame with four columns:
\itemize{
\item $id - row of sequences
\item $start - start of match (NA if no match)
\item $end - end of match (NA if no match)
\item $n_matches - number of matches (NA if no match)
}
}
\description{
Returns start end end of the primer.  If there are more than one matches, return the first match and the number of matches
}
\examples{
get_primer_position("ATT",c("ATTTTCGGG", "AGTTTCGGG"), orientation="fwd", mismatches=0)
}
