% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gg_bar_discrete}
\alias{gg_bar_discrete}
\title{Do a simple barplot}
\usage{
gg_bar_discrete(df, x, y, fill, xlimits, ymin, ymax, xlab, ylab,
  bar_labels = FALSE, flip = FALSE, legend = FALSE,
  legend_colors = c(none = "black", legend_label = ""))
}
\arguments{
\item{df}{Data frame}

\item{x}{Character - Which column use for x axis (quoted variable), should correspond to a \strong{character column}}

\item{y}{Character - Which column use for y axis (quoted variable), should correspond to a \strong{numeric column}}

\item{fill}{Character - Which column use for the slices of the bar plot (quoted variable), should correspond to a \strong{character column}}

\item{bar_labels}{TRUE or FALSE - to add the label inside the graph itself to the bar}

\item{flip}{TRUE or FALSE - to flip to horizontal mode}

\item{legend}{TRUE or FALSE - if TRUE the  two following items must be passed to the function, if FALSE then automatic legends}

\item{legend_colors}{List - Correspondance between fill categories and colors}

\item{legend_label}{Character - Title of the legend}
}
\value{
A ggplot2 object
}
\description{
Do a simple barplot
}
\examples{
p <- dv_bar_discrete (cultures_division,x="division", y="n_strains", fill="domain",
                     ymin=0, ymax=400, xlab="Division", ylab="Number of strains",
                     bar_labels=TRUE, flip=TRUE,
                     legend=TRUE,
                     legend_colors = c("deposit" = "yellow", "collaborator" = "blue",
                     "public" = "red", "education" = "green", "private"="black"),
                     legend_label = "order type")
}
