% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gg_boxplot}
\alias{gg_boxplot}
\title{Plot a box plot}
\usage{
gg_boxplot(df, y, factor)
}
\arguments{
\item{df}{Data frame - data to be plotted}

\item{y}{Character - Which column use for y axis (quoted variable), should correspond to a \strong{numeric column}}

\item{factor}{Character - Which column use for x axis (quoted variable), should correspond to a \strong{character column}}
}
\value{
a ggplot object
}
\description{
Plot a box plot
}
