% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gg_density}
\alias{gg_density}
\title{Plot simple density plot to compare factors}
\usage{
gg_density(df, x, factor)
}
\arguments{
\item{df}{Data frame - data to be plotted}

\item{x}{Character - Column corresponding x axis (quoted variable), should correspond to a \strong{numeric column}}

\item{factor}{Character - Factor for the different densities}
}
\value{
a ggplot object
}
\description{
Plot simple density plot to compare factors
}
