% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gg_hist}
\alias{gg_hist}
\title{Plot simple histogram}
\usage{
gg_hist(df, x, bin = 0.2)
}
\arguments{
\item{df}{Data frame - data to be plotted}

\item{x}{Character - Which column use for x axis (quoted variable), should correspond to a \strong{numeric column}}

\item{bin}{Numeric - Size of the bins (default = 0.2)}
}
\value{
a ggplot object
}
\description{
Plot simple histogram
}
