% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{kmer}
\alias{kmer}
\title{Get all kmers in a sequence}
\usage{
kmer(seq, kmer_width = 15)
}
\arguments{
\item{seq}{Character, the sequence to aanalyze}

\item{kmer_width}{kmer size}
}
\value{
A data frame
}
\description{
Helper function. The input is a set of sequences in a df and it returne a data frame with all kmer in column "sequence".

Note: Remove duplicates.
}
\examples{
kmer("CATATGCTTGTCTCAAAGTTAAGCCA", kmer_width = 8)
}
