% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{kmer_set}
\alias{kmer_set}
\title{Get all kmers in a set of sequence}
\usage{
kmer_set(seq_name, sequence, kmer_width = 8)
}
\arguments{
\item{seq_name}{Vector containing the sequence names (eg accession)}

\item{sequence}{Vector containing the actual sequences}

\item{kmer_width}{kmer size}
}
\value{
A data frame with 4 columns
}
\description{
From a set of sequences with name provided, it returns a data frame with the kmer common to
all sequences from the list.  It contains 4 columns: sequence, ref_sequence, start, end.

The start and end correspond to the position of the kmers on the reference sequences.
}
\section{To do}{

This may break down if the same kmer is found twice...
}

\examples{

}
