% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceano.R
\name{lat_long_dec}
\alias{lat_long_dec}
\title{Convert lat and long to decimal}
\usage{
lat_long_dec(degree, minute, orientation, coordinate_type)
}
\arguments{
\item{degree}{integer}

\item{minute}{integer}

\item{orientation}{"E" or "W" for long, "N" or "S" for lat, for the equitor can be anything}

\item{coordinate_type}{"longitude" or "latitude"}
}
\value{
Decimal coordinates or NA if coordinate_type is wrong
}
\description{
Convert lat and long to decimal
}
\examples{
lat_dec <- lat_long_dec(10, 30, "N", "latitude")
}
