% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceano.R
\name{lat_long_string_dec}
\alias{lat_long_string_dec}
\title{Convert lat and long to decimal}
\usage{
lat_long_string_dec(x)
}
\arguments{
\item{x}{character of the form 34° 52' 49"S    or 173° 17' 53"E
Must use purrr::map_dbl for mutate}
}
\value{
Decimal coordinates or NA if no coordinate
}
\description{
Convert lat and long to decimal
}
\examples{
lat_dec <- lat_long_string_dec("34° 52' 49\"S")
}
