% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{latex_fix_bibfile}
\alias{latex_fix_bibfile}
\title{Fix bib library file created by Mendeley}
\usage{
latex_fix_bibfile(
  filename = "C:/Daniel/Paper reprints Bibtex/library.bib",
  dv_bold = FALSE
)
}
\value{
TRUE if successful
}
\description{
Does the following :
\itemize{
\item Transforms <i>...</i> tag to textit
\item Replace "<="  (found in the title of some papers) by Tex symbol (leq)
\item Put Vaulot in bold -  if dv_bold=TRUE
}

Note: the bib file must be exported using the "Escape Latex character" setting OFF
}
\examples{
latex_fix_bibfile(dv_bold=TRUE)
}
