% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_distribution}
\alias{map_distribution}
\title{Draw distribution map from data frame}
\usage{
map_distribution(
  df,
  z_limits = c(0, 100),
  z_breaks = c(10, 25, 50, 75, 100),
  z_min = 1,
  color_not_present = "blue",
  size_not_present = 1,
  base_size = 14,
  map_title = "",
  map_tag = "",
  legend_position = c(0.15, 0.25),
  legend_title = "\%"
)
}
\arguments{
\item{df}{Dataframe having at least 3 columns z for the quantity to be mapped, longitude, latitude}

\item{z_limits}{Limits for the legend}

\item{z_breaks}{Breaks for the legend}

\item{z_min}{Lower limit below which the location is represented by a cross}

\item{color_not_present}{Color used to display points where the z is below z_min}

\item{base_size}{Font size for the theme (can increase to make axes legends and titles bigger)}

\item{map_title}{Title of the map}

\item{map_tag}{Tag ("A", "B") to be displayed at top right}

\item{legend_position}{Position of the legend}

\item{legend_title}{Title of the legend}
}
\value{
A map as a ggplot object
}
\description{
Draw distribution map from data frame
}
\examples{
map_distribution(df)
}
