% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_get_world}
\alias{map_get_world}
\title{Get the world map for the rworldmap package}
\usage{
map_get_world(resolution = "coarse")
}
\arguments{
\item{resolution}{"coarse" for global maps, "low" for regional maps}
}
\value{
Data frame with world map
}
\description{
Get the world map for the rworldmap package
}
