% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_leaflet}
\alias{map_leaflet}
\title{Map with leaflet package (OpenStreetMap)}
\usage{
map_leaflet(df, lng_center = 0, lat_center = 0, zoom = 3,
  width = 500, height = 500)
}
\arguments{
\item{df}{Dataframe having at least 3 columns longitude, latitude,label for label to be plotted}

\item{lng_center}{longitude of the center of the map}

\item{lat_center}{latitude of the center of the map}

\item{zoom}{map zoom, an integer from 3 (continent) to 21 (building), default value 10 (city). "auto" automatically determines the zoom for bounding box specifications, and is defaulted to 10 with center/zoom specifications. maps of the whole world currently not supported.}

\item{width}{width in pixels}

\item{height}{height in pixels}
}
\value{
A map as a leaflet object
}
\description{
Map with leaflet package (OpenStreetMap)
}
\examples{
map_leaflet(df, lng_center=0, lat_center=45, zoom=4)
}
