% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_US}
\alias{map_US}
\title{Background US map using the maps package to extract a specific country}
\usage{
map_US(color_continents = "grey80", color_borders = "white")
}
\arguments{
\item{color_continents}{color for the continents (grey80 by default)}

\item{color_borders}{color for the borders (white by default)}
}
\value{
A map as a ggplot object
}
\description{
Background US map using the maps package to extract a specific country
}
\examples{
map_US() + geom_point(data=cultures_one_year, aes(x=Longitude, y=Latitude), fill="blue", size=2, shape=21)
}
