% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_world_google}
\alias{map_world_google}
\title{Background map using the ggmap package}
\usage{
map_world_google(location = c(0, 0), zoom = 3, maptype = "satellite")
}
\arguments{
\item{location}{an address, longitude/latitude pair (in that order), or left/bottom/right/top bounding box}

\item{zoom}{map zoom, an integer from 3 (continent) to 21 (building), default value 10 (city). "auto" automatically determines the zoom for bounding box specifications, and is defaulted to 10 with center/zoom specifications. maps of the whole world currently not supported.}

\item{maptype}{character string providing map theme. options available are "terrain", "terrain-background", "satellite", "roadmap", and "hybrid" (google maps)}
}
\value{
A map as a ggplot object
}
\description{
Background map using the ggmap package
}
\examples{
map_world_google() + geom_point(data=cultures_one_year, aes(x=Longitude, y=Latitude), fill="blue", size=2, shape=21)
map_world_google(location="Singapore", zoom=11 ,maptype="hybrid")
map_world_google(location="Paris", zoom="auto", maptype="terrain")
}
