% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_zoom_europe}
\alias{map_zoom_europe}
\title{Zoom a map on Europe}
\usage{
map_zoom_europe(map)
}
\arguments{
\item{map}{Map to be zoomed}
}
\value{
A map as a ggplot object
}
\description{
Zoom a map on Europe
}
\examples{
map_zoom_europe(map)
}
