% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapr2.R
\name{metapr2_export_asv}
\alias{metapr2_export_asv}
\title{Exports the metapr2 database}
\usage{
metapr2_export_asv(taxo_level = kingdom, taxo_name = "Eukaryota",
  boot_level = class_boot, boot_min = 0,
  directory = "C:/daniel.vaulot@gmail.com/Databases/_metaPR2/export/",
  dataset_id_selected = c(1:100), export_xls = FALSE,
  export_phyloseq = FALSE, taxonomy_full = TRUE)
}
\arguments{
\item{taxo_level}{The taxonomic level for selection (do not quote), e.g. class or genus}

\item{taxo_name}{The name of the taxonomic level selected, e.g. Chlorophyta}

\item{boot_level}{The taxonomic level for bootstrap filtering (do not quote), e.g. class_boot or genus_boot}

\item{boot_min}{Minimum bootstrap value at the class level, 0 if you want to get all asvs}

\item{directory}{Directory where the files are saved (finish with /)}

\item{dataset_id_selected}{Integer vector, e.g. 23 or c(21, 23) or 21:23}

\item{export_xls}{If TRUE, an xlsx file is produced containing the final data frame}

\item{export_phyloseq}{If TRUE, a phyloseq file is produced}

\item{taxonomy_full}{If TRUE, the fasta file contains the full taxonomy (8 levels), if false only contains the species}
}
\value{
Alist with three elements that be accessed
df=asv_set ( data frame in long form)
ps=phyloseq_asv (phyloseq object)
fasta=asv_fasta (data frame with the sequences)
}
\description{
Exports a range of file and format.
\itemize{
\item fasta file with the full taxonomy or just the genus level
\item excel file with the full table of the asv and metadata
\item phyloseq file (better when only selecting a single data set)
}

Returns a list with 3 elements
\itemize{
\item df = dataframe with all the asv and the stations and read numbers
\item ps = phyloseq object
\item fasta = df with 2 columns seq_name and sequence
}
}
\examples{
# Export all the asv in a single fasta
  metapr2_export_asv()

# Export as specific data set as a phyloseq file
  metapr2_export_asv(dataset_id_selected = 23, export_phyloseq = TRUE)

# Export a specific genus as a fasta file and an excel file
  asv_set <- metapr2_export_asv(taxo_level = genus, taxo_name="Pseudohaptolina",
                                taxonomy_full= FALSE, boot_min = 100, export_xls = TRUE)
}
