% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapr2.R
\name{metapr2_export_datasets}
\alias{metapr2_export_datasets}
\title{Exports the metapr2_datasets table}
\usage{
metapr2_export_datasets()
}
\value{
A data frame
}
\description{
Exports a data frame with all the available data sets
}
\examples{
# Export all datasets
  datasets <- metapr2_export_datasets()
}
