% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapr2.R
\name{metapr2_export_metadata}
\alias{metapr2_export_metadata}
\title{Exports the metapr2_metadata table}
\usage{
metapr2_export_metadata()
}
\value{
A data frame
}
\description{
Exports a data frame with all the available metadata
}
\examples{
# Export all datasets
  metadata <- metapr2_export_metadata()
}
