% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapr2.R
\name{metapr2_export_qs}
\alias{metapr2_export_qs}
\title{Exports the metapr2 database in qs file for shiny app}
\usage{
metapr2_export_qs(set_type = "public", directory = "data/")
}
\arguments{
\item{set_type}{"public" (41 sets), "basic" (5 sets only), "all"}

\item{directory}{Directory where the files are saved (finish with /)}
}
\value{
TRUE if successful
}
\description{
Exports a range of file and format.
\itemize{
\item fasta file with the full taxonomy or just the genus level
\item excel file with the full table of the asv and metadata
\item phyloseq file (better when only selecting a single data set)
\item list of samples with metadata
}

Returns a list with 4 elements
\itemize{
\item df = dataframe with all the asv and the stations and read numbers
\item ps = phyloseq object  - if phyloseq = TRUE
\item fasta = df with 2 columns seq_name and sequence
\item samples= sample_list (list of sammples with metadata)
}

NOTE: if all export_long_xls, export_wide_xls, export_phyloseq are false, abundances are not exported
}
\examples{
# Export public data set

  metapr2_export_qs("public")

}
