% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapr2.R
\name{metapr2_export_reads_total}
\alias{metapr2_export_reads_total}
\title{Exports the total number of reads in each dataset}
\usage{
metapr2_export_reads_total(
  directory = "C:/daniel.vaulot@gmail.com/Databases/_metaPR2/export/",
  dataset_id_selected = c(1:500)
)
}
\arguments{
\item{directory}{Directory where the files are saved (finish with /)}

\item{dataset_id_selected}{Integer vector, e.g. 23 or c(21, 23) or 21:23}
}
\value{
A data frame with 2 columns:
\itemize{
\item file_code
\item reads_total
}
}
\description{
Exports  the total number of reads for each file_code

Data are save in an excel file
}
\examples{
# Export data for all datasets
  df <- metapr2_export_reads_total()
}
