% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{pcr_sequences}
\alias{pcr_sequences}
\title{In silico amplification}
\usage{
pcr_sequences(fwd_seq, rev_seq, target_seq, mismatches = 0)
}
\arguments{
\item{fwd_seq}{forward primer sequence}

\item{rev_seq}{reverse primer sequence}

\item{target_seq}{target sequences (vector)}

\item{mismatches}{Number of mismatches allowed}
}
\value{
vector with the amplicon or NA if no matches
}
\description{
Returns start end end of the primer.  If there are more than one matches, return the first match and the number of matches
}
\examples{
pcr_sequences("ATT","CCC", c("ATTTTCGGG", "AGTTTCGGG"), mismatches=0)
}
