% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{phyloseq_filter_abundant_taxa}
\alias{phyloseq_filter_abundant_taxa}
\title{Filter a phyloseq table keeping only abundant taxa}
\usage{
phyloseq_filter_abundant_taxa(ps, fraction_min = 0.1)
}
\arguments{
\item{ps}{phyloseq object}

\item{fraction_min}{minimum fraction that an OTU must represent in any given samples (not overall!) - 0.10 corresponds to 10%}
}
\value{
ps object with only abundant taxa
}
\description{
Filter a phyloseq table keeping only abundant taxa
}
\examples{
# Will return the otus that more that represent more than 10\% of total in any given sample. It also normalize the phyloseq to the median of the remaining OTUs
ps_abundant <- phyloseq_filter_abundant_taxa(ps)
# Idem but with lower threshold (5\%)
ps_abundant <- phyloseq_filter_abundant_taxa(ps, 0.05)
}
