% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{phyloseq_filter_autotrophic_taxa}
\alias{phyloseq_filter_autotrophic_taxa}
\title{Filter a phyloseq table keeping only autotrophic taxa}
\usage{
phyloseq_filter_autotrophic_taxa(ps)
}
\arguments{
\item{ps}{phyloseq object}
}
\value{
ps object with only autotrophs taxa
}
\description{
Filter a phyloseq table keeping only autotrophic taxa
}
\examples{
ps_auto <- phyloseq_filter_autotrophic_taxa(ps)
}
