% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{phyloseq_import_mothur}
\alias{phyloseq_import_mothur}
\title{Create a phyloseq file from mothur database file}
\usage{
phyloseq_import_mothur(file_mothur, file_samples, n_taxo_levels)
}
\arguments{
\item{file_mothur}{File created by mothur using the database function}

\item{file_samples}{A tsv file with the sample metadata.  There must be a column "sample_id" with the ID of the samples which should exactly matching the names of the columns in the mothur database file}

\item{n_taxo_levels}{Number of taxonomic levels (e.g. 8 if using PR2)}
}
\value{
TRUE if no problem
}
\description{
Write out 3 files:
\itemize{
\item phyloseq object as a rds file: extension .rds
\item fasta sequences (unaligned): extension .fas
\item fasta sequences (aligned): extension .aligned.fas
}
}
\examples{
phyloseq_import_mothur("mothur.database", "mothur_samples.tsv", 6)
}
