% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{phyloseq_long_treemap}
\alias{phyloseq_long_treemap}
\title{Do a treemap based on the long version of a phyloseq file}
\usage{
phyloseq_long_treemap(df, group1, group2, title)
}
\arguments{
\item{df}{Data frame obtained from a phyloseq file using the function phyloseq_transform_to_long}

\item{group1}{first grouping level (do not quote)}

\item{group2}{second grouping level (do not quote)}

\item{title}{Title for the treemap}
}
\value{
Plot the treemap
Dataframe with the summarized data
}
\description{
Plot the treemaps and returns a df with the summary of the data
}
\examples{
df <- phyloseq_long_treemap(phyloseq_long, division, class, "Singapore strait")
}
