% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{phyloseq_normalize_median}
\alias{phyloseq_normalize_median}
\title{Normalize to the median number of reads}
\usage{
phyloseq_normalize_median(ps)
}
\arguments{
\item{ps}{phyloseq object}
}
\value{
ps object normalized
}
\description{
Normalize to the median number of reads
}
\examples{
ps_norm <- phyloseq_normalize_median(ps)
}
