% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{phyloseq_normalize_percent}
\alias{phyloseq_normalize_percent}
\title{Normalize as percent of total reads}
\usage{
phyloseq_normalize_percent(ps)
}
\arguments{
\item{ps}{phyloseq object}
}
\value{
ps object normalized
}
\description{
Normalize as percent of total reads
}
\examples{
ps_percent <- phyloseq_normalize_percent(ps)
}
