% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_buid_taxons}
\alias{pr2_buid_taxons}
\title{Build taxon table (long format)}
\usage{
pr2_buid_taxons(pr2_taxonomy, pr2.env, add_old_fields = FALSE)
}
\arguments{
\item{pr2_taxonomy}{dataframe - should contain eight columns kingdom -> species (this can be modified by using the pr2.env list)}

\item{pr2.env}{list containing the taxon levels}

\item{add_old_fields}{logical - If add_old_fields = TRUE then the following columns are appended: old_species_id, taxon_edited_version, taxon_edited_by,  taxon_removed_version, taxon_remark,  taxon_reference}
}
\value{
Data frame pr2_taxons with columns "taxon_id" , "taxon_name", "taxon_level", "parent_id", "old_species_id",
"taxon_edited_version", "taxon_edited_by", "taxon_removed_version" ,"taxon_remark", "taxon_reference"
}
\description{
This constructs a taxon table (long format) from a taxonomy table (wide format)
If add_old_fields = TRUE then the following columns are appended
old_species_id, taxon_edited_version, taxon_edited_by,  taxon_removed_version, taxon_remark,  taxon_reference
}
\note{
The input data frame does not need to be the whole pr2 taxo, can be taxo extracted from an Excel file.
}
\examples{
pr2_taxons <- pr2_buid_taxons(pr2_taxonomy)
}
