% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_colors_read}
\alias{pr2_colors_read}
\title{Reads the PR2 color into a data frame}
\usage{
pr2_colors_read()
}
\value{
A data frame with all the columns from pr2_colors.
}
\description{
The database access codes is obtained with the function db_info("pr2_google").

\strong{Important}: This gets the current taxonomic  database and does not include taxa that have been removed.
}
\examples{
pr2_colors <- pr2_colors_read()
}
