% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_export}
\alias{pr2_export}
\title{Export the PR2 database (one file)}
\usage{
pr2_export(
  pr2_select,
  file_name,
  file_type = "fasta",
  file_format = "fasta_taxo_long",
  taxo_levels_number = 9
)
}
\arguments{
\item{pr2_select}{data frame - the pr2 database or an extract of the pr2 database}

\item{file_name}{character - full path of file where to save}

\item{file_type}{character - one of "fasta", "taxo", "metadata", "merged", "merged_excel"}

\item{file_format}{character - one of "UTAX", "fasta_taxo_short","fasta_taxo_long","dada2", "dada2_species","mothur"}
}
\value{
Write the files in compressed format (.gz)
}
\description{
This will save the pr2 database in variety of format. The files are compressed as .gz
\itemize{
\item "fasta" - the description line is taylored for different applications (usearch, mothur, dada2, blast). for the application dada2 2 kinds of files can be produced one for otu assignement and one for species
\item "taxo" - the file is saved with the mothur format
\item "metadata" - only the metadata
\item "merged" - the whole database in a single file
}

Notes
\itemize{
\item The data are NOT filtered in any way, this should be done before saving
\item The following fields are needed to create the sequence labels : pr2_sample_type, gb_strain, gb_clone, gb_specimen_voucher
}
}
\examples{
pr2_export(pr2, "C:/Daniel/myfile.fas", "fasta", "mothur", 9)
}
