% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_export_all}
\alias{pr2_export_all}
\title{Export the PR2 database (all files)}
\usage{
pr2_export_all(pr2.env)
}
\arguments{
\item{pr2.env}{list containing the name of files}
}
\value{
Write the all the files in compressed format (.gz)
}
\description{
This will save the pr2 database in all the format necessary for the repository on GitHub. It calls \code{link{pr2_export}}
}
\examples{
pr2_export_all(pr2.env)
}
