% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_export_duckdb}
\alias{pr2_export_duckdb}
\title{Export the PR2 database to a Duckdb databasee}
\usage{
pr2_export_duckdb(file_name)
}
\arguments{
\item{file_name}{character - full name where to save}
}
\value{
TRUE is successful
}
\description{
This will save the pr2 database in Duckdb database as tables
\itemize{
\item pr2_main
\item pr2_sequences
\item pr2_metadata
\item pr2_taxo
\item pr2_assign_bayes
\item pr2_traits
\item pr2_countries
\item pr2_primers
\item pr2_primer_sets
\item worms_taxonomy
\item eukribo_v2
\item pr2_taxonomy_5.0_8_levels
NOTE: duckdb file size is same as SQLite but cannot be read with DBeaver
}
}
\examples{
pr2_export_duckdb("Mydirectory/pr2.duckdb")
}
