% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_export_emu}
\alias{pr2_export_emu}
\title{Export the PR2 database to construct database for emu}
\usage{
pr2_export_emu(
  pr2_directory = "C:/daniel.vaulot@gmail.com/Databases/_PR2/",
  version = "5.0.1"
)
}
\arguments{
\item{pr2_directory}{Directory above the versions directory}

\item{version}{PR2 version as character (e.g. 5.0.0)}
}
\value{
TRUE if successful
}
\description{
Three files are exported
}
\examples{
pr2 <- pr2_export_emu()
}
