% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_export_sqlite}
\alias{pr2_export_sqlite}
\title{Export the PR2 database to a SQLite file}
\usage{
pr2_export_sqlite(file_name)
}
\arguments{
\item{file_name}{character - full path of file where to save}
}
\value{
TRUE is succesful
}
\description{
This will save the pr2 database in SQLite databas as 4 tables (pr2_main, pr2_sequences, pr2_metadata, pr2_taxo)
}
\examples{
pr2_export_sqlite("pr2_local.sqlite")
}
