% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_sequence_label}
\alias{pr2_sequence_label}
\title{Create a simple label for a sequence}
\usage{
pr2_sequence_label(sample_type, strain, clone, specimen_voucher)
}
\arguments{
\item{sample_type}{character - one of "culture", "environmental","isolate", "species"}

\item{strain}{character - strain name}

\item{clone}{character - clone name}

\item{specimen_voucher}{character - voucher}
}
\value{
character
}
\description{
Create a label to be used as header a fasta file based on the metadata from the table \emph{pr2_metadata}. Only apply to a single character not to a vector.
}
\examples{
fasta_label <- pr2_sequence_label(sample_type="culture",strain="RCC115")
}
