% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_sequence_reassign}
\alias{pr2_sequence_reassign}
\title{Reassign pr2 sequences}
\usage{
pr2_sequence_reassign(pr2_update_old)
}
\arguments{
\item{pr2_update_old}{data frame - should have at least two columns:
\itemize{
\item genbank_accession
\item species
}}
}
\value{
The input data frame is returned with two extra columns
\itemize{
\item query : the executed query
\item query_result : whether the query was successful
}
}
\description{
Build a query to reassign genbank entries to a new species in the table \emph{pr2_main}

Two other fields of \emph{pr2_main} are also written using the pr2.env which is loaded with the file PR2_init.R
\itemize{
\item edited_version : pr2.env$version
\item edited_by : pr2.env$editor
}
}
\examples{
genbank_list_result <- pr2_sequence_reassign(genbank_list)
}
