% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_taxo_full_read}
\alias{pr2_taxo_full_read}
\title{Reads the PR2 taxonomic database into a data frame with WORMS, GBIF and traits information}
\usage{
pr2_taxo_full_read(taxo_levels_number = 9, trait_types = c("mixoplankton"))
}
\value{
A data frame
}
\description{
The database access codes is obtained with the function db_info("pr2_google").

\strong{Important}: This gets the current taxonomic  database and does not include taxa that have been removed.
}
\examples{
pr2_taxo <- pr2_taxo_read(9, c("mixoplankton"))
}
