% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_taxo_list}
\alias{pr2_taxo_list}
\title{Build a list of taxa}
\usage{
pr2_taxo_list(pr2, level_select)
}
\arguments{
\item{pr2}{\itemize{
\item data frame - whole or filtered pr2 data base
}}

\item{level_select}{\itemize{
\item taxonomic rank
}}
}
\value{
A character vector
}
\description{
Construct a list of taxa from the pr2 database
}
\examples{
pr2_taxo_list(pr2, "class")
}
