% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_taxo_read}
\alias{pr2_taxo_read}
\title{Reads the PR2 taxonomic database into a data frame}
\usage{
pr2_taxo_read()
}
\value{
A data frame with all the columns from pr2_taxo.
}
\description{
The database access codes is obtained with the function db_info("pr2_local").

\strong{Important}: This gets the current taxonomic  database and does not include taxa that have been removed.
}
\examples{
pr2_taxo <- pr2_taxo_read()
}
