% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_taxo_update}
\alias{pr2_taxo_update}
\title{Update pr2 taxonomy}
\usage{
pr2_taxo_update(pr2_taxo, method = "update")
}
\arguments{
\item{pr2_taxo}{data frame - Should have at least 8 columns:
\itemize{
\item kingdom
\item ...
\item species
}}

\item{method}{character - \code{method} = "update" update existing list of species, = "add"  add a list of new species}
}
\value{
The input data frame is returned with two extra columns
\itemize{
\item query : the executed query
\item query_result : whether the query was successful
}
}
\description{
Build a query to modify the table \emph{taxo} to change the fields kingdom to genus for an existing species or to add a new species.

Two other fields of \emph{taxo} are also written using pr2.env environment loaded in PR2_init.R
\itemize{
\item taxo_edited_version : pr2.env$version
\item taxo_edited_by : pr2.env$editor
}

In addition the genus path is updated for all species belonging to the same gneus so that all species
from a given genus have the same upward taxonomy.  For example if \emph{Mypico marinus} is reassigned to the Mamiellophyceae
all species from the genus \emph{Mypico} should reassigned to Mamiellophyceae.
}
\examples{
taxo_list_result <- pr2_taxo_update(taxo_list)
}
