% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_taxo_X}
\alias{pr2_taxo_X}
\title{Add _X for taxon that are repeated on same line}
\usage{
pr2_taxo_X(pr2_taxo, taxo_levels_number = 9)
}
\arguments{
\item{pr2_taxo}{Data frame containing columns from kingdom to species (or else see the next parameter).  It is not necessary that the columns be ordered.}

\item{taxo_levels}{A vector of ordered taxonomic ranks such as c("kingdom", "supergroup", "division", "class", "order", "family", "genus", "species")}
}
\value{
A corrected data frame
}
\description{
If more than 2 ranks on a given line have the same name or is empty, rename with Mytaxon_X, Mytaxon_XX etc.. up to the species name which becomes Mytaxon_XX_sp.
}
\examples{
pr2_taxo_clean <- pr2_taxo_X(pr2_taxo, 9)
}
