% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_traits_merge}
\alias{pr2_traits_merge}
\title{Merge the PR2 traits table with taxo table into a data frame}
\usage{
pr2_traits_merge(
  trait_types = c("mixoplankton"),
  taxo_levels_number = 9,
  debug = FALSE,
  debug_file = here::here("tests", "testthat", "output",
    "pr2_traits_species_duplicated.xlsx")
)
}
\arguments{
\item{trait_types}{character vector - e.g. "mixotrophy"}

\item{taxo_levels_number}{number - number of taxonomical levels (8 or 9)}

\item{debug}{logical - Save files of duplicates}

\item{debug_file}{character - Name for files of duplicates}
}
\value{
A data frame
}
\description{
The database access codes is obtained with the function db_info("pr2_google").

\strong{Important}: This gets the current taxonomic  database and does not include taxa that have been removed.
}
\examples{
pr2_taxo_traits <- pr2_traits_merge(trait_types = c("mixotrophy", "ecological_function"), taxo_levels_number = 9, debug = TRUE)
}
