% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_traits_read}
\alias{pr2_traits_read}
\title{Reads the PR2 traits table into a data frame}
\usage{
pr2_traits_read()
}
\value{
A data frame
}
\description{
The database access codes is obtained with the function db_info("pr2_google").

\strong{Important}: This gets the current taxonomic  database and does not include taxa that have been removed.
}
\examples{
pr2_traits <- pr2_traits_read()
}
