% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_analysis.R
\name{pr2_treemap}
\alias{pr2_treemap}
\title{Draw a simple treemap from pr2.}
\usage{
pr2_treemap(pr2, level1, level2)
}
\arguments{
\item{pr2}{data frame - can be the whole database or a filtered version of it.
It needs at three columns, 2 taxonomy columns and one for example with sequences}

\item{level1}{unquoted variable - lower grouping level}

\item{level2}{unquoted variable - higher grouping level}
}
\description{
Can also be used for any data frame with same taxo structure
}
\examples{
# Will plot at the order and family levels
pr2_treemap (pr2_filtered, species, order)
}
