% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc.R
\name{rcc_customers}
\alias{rcc_customers}
\title{Format "customers.csv" for import into rcc database}
\usage{
rcc_customers(customers_file)
}
\arguments{
\item{customers_file}{Name of file "customers.csv" exported from the web site}
}
\value{
Data frame ready for import into rcc database
}
\description{
This function reads the file "customers.csv" exported from the web site
and import it into the rcc database
}
\examples{
df <- rcc_customers("customers.csv")
}
