% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc.R
\name{rcc_genbank}
\alias{rcc_genbank}
\title{Prepare list of GenBank entries to include in rcc database}
\usage{
rcc_genbank(genbank_file)
}
\arguments{
\item{genbank_file}{Name of excel file exported from Geneious}
}
\value{
Data frame ready for import into rcc database. .
}
\description{
Input is an Excel file exported from Geneious with three additional columns
\itemize{
\item rcc_id
\item gene_location
\item gene_name
}

Entries already in the rcc database are discarded
}
\examples{
df <- rcc_genbank("RCC from Genbank 2017 12 26.xlsx")
}
