% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc.R
\name{rcc_order_details}
\alias{rcc_order_details}
\title{Format "order_details.csv" for import into rcc database}
\usage{
rcc_order_details(order_details_file)
}
\arguments{
\item{order_details_file}{Name of file "order_details.csv" exported from the web site}
}
\value{
Data frame ready for import into rcc database
}
\description{
This function reads the file "order_details.csv" exported from the web site
and import it into the rcc database
}
\examples{
df <- rcc_order_details("order_details.csv")
}
