% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc.R
\name{rcc_orders}
\alias{rcc_orders}
\title{Format "orders.csv" for import into rcc database}
\usage{
rcc_orders(orders_file)
}
\arguments{
\item{orders_file}{Name of file "orders.csv" exported from the web site}
}
\value{
Data frame ready for import into rcc database
}
\description{
This function reads the file "orders.csv" exported from the web site
and import it into the rcc database
}
\examples{
df <- rcc_orders("orders.csv")
}
