% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{read_db_config}
\alias{read_db_config}
\title{Get the connection info from a yaml file}
\usage{
read_db_config(
  db_config_yaml = "C:/daniel.vaulot@gmail.com/Databases/google_pc.yaml"
)
}
\arguments{
\item{db_config_yaml}{The yaml file containing the info}
}
\value{
A list with the database connection info
\itemize{
\item db_type: mysql, duckdb ...
\item db_config$db_config_file: for MySQL, the location of my.cnf
\item db_config$db_groups: for MySQL
\item db_file_pr2: for duckdb
\item db_file_metapr2: for duckdb
}
}
\description{
To be used with db_connect (see the example).
The MySQL connection info are stored in a file (my.cnf). See https://github.com/r-dbi/RMySQL/issues/131
}
\examples{
db_config <- read_db_config("C:/daniel.vaulot/Databases/google_pc.yaml") # Must always use db_config
db_connect("pr2")
}
