% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{rod_read}
\alias{rod_read}
\title{Export the ROD database to data frame}
\usage{
rod_read(taxo_levels_number = 9, trait_types = c("mixoplankton"))
}
\arguments{
\item{directory}{string direcotry where to save the files}
}
\value{
Write the the files in qs format and return the rod list
}
\description{
This will save the ROD database as a dataframe.
}
\examples{
rod_export_qs("C:/Desktop/")
}
