% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{round_down_to_nearest_100}
\alias{round_down_to_nearest_100}
\title{Round a value to previous 100}
\usage{
round_down_to_nearest_100(number)
}
\value{
New value
}
\description{
Round a value to previous 100
}
\examples{
numbers <- c(150, 230, 780, 900, 1001, 0, -250)
rounded_numbers <- sapply(numbers, round_down_to_nearest_100)
}
