% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{seq_hash}
\alias{seq_hash}
\title{Sequence hash (vectorial)}
\usage{
seq_hash(seqs)
}
\arguments{
\item{seq}{vector of sequences}
}
\value{
Hash value as a vector
}
\description{
Get the hash value for a sequences
}
\examples{
seq_hash (c("ATTTTCGGG", "ATTTTTGGG"))
}
