% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{seq_reverse_complement}
\alias{seq_reverse_complement}
\title{Reverse complement string}
\usage{
seq_reverse_complement(seqs)
}
\arguments{
\item{seq}{vector of sequences}
}
\value{
Reverse complement vector of strings
}
\description{
Reverse complement a set of sequences
}
\examples{
seq_reverse_complement (c("ATTTTCGGG", "ATTTTCGGG"))
}
