% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_dviz_vgrid}
\alias{theme_dviz_vgrid}
\title{Theme - vertical grid lines only}
\usage{
theme_dviz_vgrid(font_size = 14, font_family = "")
}
\arguments{
\item{font_size}{font size}

\item{font_family}{font family}
}
\value{
Theme
}
\description{
From data viz book : http://serialmentor.com/dataviz/
}
