% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{treemap_dv}
\alias{treemap_dv}
\title{Do a simple treemap}
\usage{
treemap_dv(df, index, size, title)
}
\arguments{
\item{df}{Data frame}

\item{index}{Character vector - what variables are used for categories}

\item{size}{Character - what variable is use for size of boxes}

\item{title}{Character - Title}
}
\value{
A treemap object and plots the treemap
}
\description{
Do a simple treemap
}
\examples{
treemap_dv(dada2_species, c("Division", "Class"),"n_seq","Dada2" )
}
