% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unfill}
\alias{unfill}
\title{Remove repeated values in vector}
\usage{
unfill(x)
}
\value{
Vector with repeated values replaced by NA
}
\description{
Repeated values are replaced by NA. This is useful for formatting tables for papers.
}
\examples{
x <- c("A","A","A","B","B","C","C")
unfill(x)
\dontrun{
mutate(df, across(c("division", "class"), ~ unfill(.x)))
}
}
