% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{XStringSet_to_df}
\alias{XStringSet_to_df}
\title{Transforms a DNA or AA String set into a data frame}
\usage{
XStringSet_to_df(seq_set)
}
\arguments{
\item{seq_set}{DNAStringSet or AA}
}
\value{
A data frame
}
\description{
Very simple helper function. The data frame has two columns seq_name and sequence
}
\examples{
df.fasta <- DNAStrinSet_to_df(DNAset_fasta)
}
