% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Establish a connection database}
\usage{
db_connect(db_info)
}
\arguments{
\item{db_info}{list(default.file= ,groups= ,dbname= )}
}
\value{
A database connection
}
\description{
All connections parameters are stored into a my.conf file
}
\examples{
pr2_con <- db_connect(db_info("pr2_local"))
test_con <- db_connect(list(default.file="my.cnf", groups="root" , dbname= "test"))
}
