% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{phyloseq_long_treemap}
\alias{phyloseq_long_treemap}
\title{Do a treemap based on the long version of a phyloseq file}
\usage{
phyloseq_long_treemap(df, group1, group2, title, colors = NULL)
}
\arguments{
\item{df}{Data frame obtained from a phyloseq file using the function phyloseq_transform_to_long}

\item{group1}{first grouping level (do not quote)}

\item{group2}{second grouping level (do not quote)}

\item{title}{Title for the treemap}

\item{colors}{If NULL then the default viridis palette is used.  If named vectors then the colors}
}
\value{
Plot the treemap
Returns a list made of 2 elements
\itemize{
\item gg = the treemap as ggplot
\item df =  a dataframe with the summarized data
}
}
\description{
Plot the treemaps and returns a list with a ggplot and a df with the summary of the data
}
\examples{
my_list <- phyloseq_long_treemap(phyloseq_long, division, class, "Singapore strait", colors=
                                 structure(c("black", "white"),.Names="Mamiellophyceae", "Dinophyceae"))
}
