% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_export_all}
\alias{pr2_export_all}
\title{Export the PR2 database (all files)}
\usage{
pr2_export_all(
  pr2_directory = "C:/daniel.vaulot@gmail.com/Databases/_PR2/",
  version = "5.0.0",
  taxo_levels_number = 9,
  traits_used = c("mixoplankton"),
  test = FALSE
)
}
\arguments{
\item{pr2_directory}{Directory above the versions directory}

\item{version}{PR2 version as character (e.g. 5.0.0)}

\item{taxo_levels_number}{Number of taxonomic levels}

\item{traits_used}{Traits included in the database}

\item{test}{If true only 2000 sequences}
}
\value{
Write the all the files in compressed format (.gz) or xlsx format
}
\description{
This will save the pr2 database in all the format necessary for the repository on GitHub. It calls \code{link{pr2_export}}.
It also saves 4 Excel files (taxonomy, chimera, updated, unassigned)
It does not save the qs file used for the package that are generated by the data_raw script from the pr2database package
}
\examples{
pr2_export_all(version = "5.0.0", test = TRUE)
}
