% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_export_sqlite}
\alias{pr2_export_sqlite}
\title{Export the PR2 database to a SQLite file}
\usage{
pr2_export_sqlite(file_name)
}
\arguments{
\item{file_name}{character - full path of file where to save}
}
\value{
TRUE is succesful
}
\description{
This will save the pr2 database in SQLite database, the following tables
\itemize{
\item pr2_main
\item pr2_sequences
\item pr2_metadata
\item pr2_taxo
\item pr2_assign_bayes
\item pr2_traits
\item pr2_countries
\item pr2_primers
\item pr2_primer_sets
\item taxonomy_worms
\item eukribo_v2
\item pr2_taxonomy_5.0_8_levels
}
}
\examples{
pr2_export_sqlite("pr2_google.sqlite")
}
