% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2_management.R
\name{pr2_read}
\alias{pr2_read}
\title{Reads the whole PR2 database into a data frame}
\usage{
pr2_read()
}
\value{
A data frame with all the columns from pr2.
}
\description{
The database access codes is obtained with the function db_info("pr2_google").

\strong{Important}: This gets the whole database, including sequences that have been removed.  These sequences must be filtered out before export using pr2  <- pr2 \%>\% filter (is.na(removed_version)). However the taxo entries that have been removed are already removed.
}
\examples{
pr2 <- pr2_read()
}
